+++ server/src/graql/admin/ReasonerQuery.java
@@ -21,6 +21,8 @@ package grakn.core.graql.admin;
 import com.google.common.collect.ImmutableMap;
 import grakn.core.graql.answer.ConceptMap;
 import grakn.core.graql.concept.Type;
+import grakn.core.graql.internal.reasoner.rule.InferenceRule;
+import grakn.core.graql.internal.reasoner.rule.RuleUtils;
 import grakn.core.graql.query.pattern.Conjunction;
 import grakn.core.graql.query.pattern.Pattern;
 import grakn.core.graql.query.pattern.Variable;
@@ -125,6 +127,14 @@ public interface ReasonerQuery{
      */
     @CheckReturnValue
     Stream<ConceptMap> resolve();
+
+    /**
+     * reiteration might be required if rule graph contains loops with negative flux
+     * or there exists a rule which head satisfies body
+     * @return true if because of the rule graph form, the resolution of this query may require reiteration
+     */
+    @CheckReturnValue
+    boolean requiresReiteration();
 
     /**
      * Returns a var-type map local to this query. Map is cached.
